Termino de testear prueba de agregador integral

+++ agregador/src/main/java/aplicacion/dto/SolicitudDTO.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class SolicitudDTO {
-    private Integer solicitanteId;
+    private Long solicitanteId;
     private String hechoId;
     private String motivo;
 }
